requirements.txt
Authlib==0.14.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
Flask==1.1.1
furl==2.1.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
numpy==1.18.2
orderedmultidict==1.0.1
pandas==1.0.3
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2019.3
requests==2.23.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
xlrd==1.2.0
xmltodict==0.13.0